fin crud de categorias

+++ app/Http/Controllers/Admin/FamilyController.php
@@ -36,6 +36,15 @@ class FamilyController extends Controller
 
         Family::create($validated);
 
+        session()->flash('swal', [
+            'title' => "¡Familia creada!",
+            'text' => "La familia $request->name ahora está disponible",
+            'icon' => "success",
+            'draggable' => true,
+            'timer' => 1600,
+            'timerProgressBar' => true
+        ]);
+        
         return redirect()->route('admin.families.create');
     }
 
@@ -66,6 +75,14 @@ class FamilyController extends Controller
 
         $family->update($validated);
 
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => '¡Familia actualizada!',
+            'text' => "La familia $family->name ha sido actualizada correctamente",
+            'timer' => 1600,
+            'timerProgressBar' => true
+        ]);
+
         return redirect()->route('admin.families.edit', $family);
     }
 
@@ -74,7 +91,27 @@ class FamilyController extends Controller
      */
     public function destroy(Family $family)
     {
+        if ($family->categories->count() > 0) {
+            session()->flash('swal', [
+                'icon' => 'error',
+                'title' => '¡Ups!',
+                'text' => "No se puede eliminar la familia $family->name, porque tiene categorias asociadas",
+                'timer' => 1600,
+                'timerProgressBar' => true
+            ]);
+
+            return redirect()->route('admin.families.index');
+        }
+
         $family->delete();
+
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => '¡Familia eliminada!',
+            'text' => "La familia $family->name ha sido eliminada correctamente",
+            'timer' => 1600,
+            'timerProgressBar' => true
+        ]);
 
         return redirect()->route('admin.families.index');
     }
